Make the error page text more generic
The error pages may be served in response to http
errors from any application running in the cloud
platform (or if a user tries to access a URL which
doesn't match anything). So, the copy should be
as generic as possible - the user might not have
entered any data, so we shouldn't say "your 
answers have not been saved" because there might
not have been any answers to save.